+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add implementation for `\Heptacom\HeptaConnect\Storage\Base\Contract\RouteOverviewActionInterface` in `\Heptacom\HeptaConnect\Storage\ShopwareDal\Action\RouteOverview`
 - Add implementation for `\Heptacom\HeptaConnect\Storage\Base\Contract\RouteFindByTargetsAndTypeActionInterface` in `\Heptacom\HeptaConnect\Storage\ShopwareDal\Action\RouteFindByTargetsAndType`
 - Add implementation for `\Heptacom\HeptaConnect\Storage\Base\Contract\RouteGetActionInterface` in `\Heptacom\HeptaConnect\Storage\ShopwareDal\Action\RouteGet`
+- Add implementation for `\Heptacom\HeptaConnect\Storage\Base\Contract\RouteCreateActionInterface` in `\Heptacom\HeptaConnect\Storage\ShopwareDal\Action\RouteCreate`
 - Add `\Heptacom\HeptaConnect\Storage\ShopwareDal\Support\Query\QueryIterator` to simplify DBAL paginated iteration
 
 ### Changed